Moved sankey series into a subfolder.
