+++ ts/Series/DependencyWheelSeries.ts
@@ -13,7 +13,7 @@
 'use strict';
 
 import type PieSeries from './Pie/PieSeries';
-import type SankeySeries from './SankeySeries';
+import type SankeySeries from './Sankey/SankeySeries';
 import type { SeriesStatesOptions } from '../Core/Series/SeriesOptions';
 import type SVGAttributes from '../Core/Renderer/SVG/SVGAttributes';
 import type SVGElement from '../Core/Renderer/SVG/SVGElement';
